+++ C Utilities/DataStream.h
@@ -7,16 +7,16 @@
 
 typedef struct {
 	Array Data;
-	uint64 Cursor;
+	uint32 Cursor;
 	boolean IsEOF;
 } DataStream;
 
-exported DataStream* DataStream_New(uint64 allocation);
-exported void DataStream_Initialize(DataStream* dataStream, uint64 allocation);
+exported DataStream* DataStream_New(uint32 allocation);
+exported void DataStream_Initialize(DataStream* dataStream, uint32 allocation);
 exported void DataStream_Free(DataStream* self);
 exported void DataStream_Uninitialize(DataStream* self);
 
-exported void DataStream_Seek(DataStream* self, uint64 position);
+exported void DataStream_Seek(DataStream* self, uint32 position);
 
 exported void DataStream_WriteInt8(DataStream* self, int8 data);
 exported void DataStream_WriteInt16(DataStream* self, int16 data);
@@ -28,7 +28,7 @@ exported void DataStream_WriteUInt32(DataStream* self, uint32 data);
 exported void DataStream_WriteUInt64(DataStream* self, uint64 data);
 exported void DataStream_WriteFloat32(DataStream* self, float32 data);
 exported void DataStream_WriteFloat64(DataStream* self, float64 data);
-exported void DataStream_WriteBytes(DataStream* self, uint8* data, uint64 count, boolean disposeBytes);
+exported void DataStream_WriteBytes(DataStream* self, uint8* data, uint32 count, boolean disposeBytes);
 exported void DataStream_WriteArray(DataStream* self, Array* array, boolean disposeArray);
 exported void DataStream_WriteString(DataStream* self, String* string, boolean disposeString);
 
@@ -42,8 +42,8 @@ exported uint32 DataStream_ReadUInt32(DataStream* self);
 exported uint64 DataStream_ReadUInt64(DataStream* self);
 exported float32 DataStream_ReadFloat32(DataStream* self);
 exported float64 DataStream_ReadFloat64(DataStream* self);
-exported uint8* DataStream_ReadBytes(DataStream* self, uint64 count);
-exported Array* DataStream_ReadArray(DataStream* self, uint64 count);
+exported uint8* DataStream_ReadBytes(DataStream* self, uint32 count);
+exported Array* DataStream_ReadArray(DataStream* self, uint32 count);
 exported String* DataStream_ReadString(DataStream* self);
 
 #endif
